+++ GFramework.Core/Command/AbstractAsyncCommand.cs
@@ -1,5 +1,6 @@
 using GFramework.Core.Abstractions.Command;
 using GFramework.Core.Rule;
+using GFramework.Cqrs.Abstractions.Cqrs.Command;
 
 namespace GFramework.Core.Command;
 
@@ -25,4 +26,55 @@ public abstract class AbstractAsyncCommand : ContextAwareBase, IAsyncCommand
     /// </summary>
     /// <returns>表示异步操作的任务</returns>
     protected abstract Task OnExecuteAsync();
+}
+
+/// <summary>
+///     抽象异步命令基类，为需要命令输入且无返回值的异步命令提供统一执行骨架。
+/// </summary>
+/// <typeparam name="TInput">命令输入类型，必须实现 <see cref="ICommandInput" /> 接口。</typeparam>
+/// <param name="input">命令输入参数。</param>
+public abstract class AbstractAsyncCommand<TInput>(TInput input) : ContextAwareBase, IAsyncCommand
+    where TInput : ICommandInput
+{
+    /// <summary>
+    ///     执行异步命令的实现方法。
+    /// </summary>
+    /// <returns>表示异步操作的任务。</returns>
+    async Task IAsyncCommand.ExecuteAsync()
+    {
+        await OnExecuteAsync(input).ConfigureAwait(false);
+    }
+
+    /// <summary>
+    ///     定义异步执行逻辑的抽象方法，由派生类实现具体业务逻辑。
+    /// </summary>
+    /// <param name="input">命令输入参数。</param>
+    /// <returns>表示异步操作的任务。</returns>
+    protected abstract Task OnExecuteAsync(TInput input);
+}
+
+/// <summary>
+///     抽象异步命令基类，为需要命令输入且返回结果的异步命令提供统一执行骨架。
+/// </summary>
+/// <typeparam name="TInput">命令输入类型，必须实现 <see cref="ICommandInput" /> 接口。</typeparam>
+/// <typeparam name="TResult">命令执行结果类型。</typeparam>
+/// <param name="input">命令输入参数。</param>
+public abstract class AbstractAsyncCommand<TInput, TResult>(TInput input) : ContextAwareBase, IAsyncCommand<TResult>
+    where TInput : ICommandInput
+{
+    /// <summary>
+    ///     执行异步命令并返回结果的实现方法。
+    /// </summary>
+    /// <returns>表示异步操作且包含结果的任务。</returns>
+    async Task<TResult> IAsyncCommand<TResult>.ExecuteAsync()
+    {
+        return await OnExecuteAsync(input).ConfigureAwait(false);
+    }
+
+    /// <summary>
+    ///     定义异步执行逻辑的抽象方法，由派生类实现具体业务逻辑并返回结果。
+    /// </summary>
+    /// <param name="input">命令输入参数。</param>
+    /// <returns>表示异步操作且包含结果的任务。</returns>
+    protected abstract Task<TResult> OnExecuteAsync(TInput input);
 }
