+++ GFramework.Core/Events/Event.cs
@@ -3,8 +3,8 @@
 namespace GFramework.Core.Events;
 
 /// <summary>
-///     泛型事件类，支持一个泛型参数 T 的事件注册、注销与触发。
-///     实现了 IEvent 接口以提供统一的事件操作接口。
+///     泛型事件类，支持一个泛型参数 <typeparamref name="T" /> 的事件注册、注销与触发。
+///     实现了 <see cref="IEvent" /> 接口以提供统一的事件操作接口。
 /// </summary>
 /// <typeparam name="T">事件回调函数的第一个参数类型。</typeparam>
 public class Event<T> : IEvent
@@ -16,11 +16,11 @@ public class Event<T> : IEvent
     private Action<T>? _mOnEvent = _ => { };
 
     /// <summary>
-    ///     显式实现 IEvent 接口中的 Register 方法。
-    ///     允许使用无参 Action 来订阅当前带参事件。
+    ///     显式实现 <see cref="IEvent" /> 接口中的 <c>Register</c> 方法。
+    ///     允许使用无参 <see cref="Action" /> 来订阅当前带参事件。
     /// </summary>
     /// <param name="onEvent">无参事件处理方法。</param>
-    /// <returns>IUnRegister 对象，用于稍后注销该事件监听器。</returns>
+    /// <returns><see cref="IUnRegister" /> 对象，用于稍后注销该事件监听器。</returns>
     IUnRegister IEvent.Register(Action onEvent)
     {
         return Register(Action);
@@ -35,7 +35,7 @@ public class Event<T> : IEvent
     ///     注册一个事件监听器，并返回可用于取消注册的对象。
     /// </summary>
     /// <param name="onEvent">要注册的事件处理方法。</param>
-    /// <returns>IUnRegister 对象，用于稍后注销该事件监听器。</returns>
+    /// <returns><see cref="IUnRegister" /> 对象，用于稍后注销该事件监听器。</returns>
     public IUnRegister Register(Action<T> onEvent)
     {
         _mOnEvent += onEvent;
@@ -52,7 +52,7 @@ public class Event<T> : IEvent
     }
 
     /// <summary>
-    ///     触发所有已注册的事件处理程序，并传递参数 t。
+    ///     触发所有已注册的事件处理程序，并传递参数 <paramref name="t" />。
     /// </summary>
     /// <param name="t">传递给事件处理程序的参数。</param>
     public void Trigger(T t)
@@ -61,9 +61,9 @@ public class Event<T> : IEvent
     }
 
     /// <summary>
-    ///     获取当前已注册的监听器数量
+    ///     获取当前已注册的监听器数量。
     /// </summary>
-    /// <returns>监听器数量</returns>
+    /// <returns>监听器数量。</returns>
     public int GetListenerCount()
     {
         return _mOnEvent?.GetInvocationList().Length ?? 0;
@@ -71,30 +71,30 @@ public class Event<T> : IEvent
 }
 
 /// <summary>
-///     支持两个泛型参数 T 和 TK 的事件类。
+///     支持两个泛型参数 <typeparamref name="T" /> 和 <typeparamref name="TK" /> 的事件类。
 ///     提供事件注册、注销和触发功能。
 /// </summary>
 /// <typeparam name="T">第一个参数类型。</typeparam>
-/// <typeparam name="Tk">第二个参数类型。</typeparam>
-public class Event<T, Tk> : IEvent
+/// <typeparam name="TK">第二个参数类型。</typeparam>
+public class Event<T, TK> : IEvent
 {
     /// <summary>
     ///     存储已注册的双参数事件处理委托。
     ///     默认为空操作（no-op）委托。
     /// </summary>
-    private Action<T, Tk>? _mOnEvent = (_, _) => { };
+    private Action<T, TK>? _mOnEvent = (_, _) => { };
 
     /// <summary>
-    ///     显式实现 IEvent 接口中的 Register 方法。
-    ///     允许使用无参 Action 来订阅当前带参事件。
+    ///     显式实现 <see cref="IEvent" /> 接口中的 <c>Register</c> 方法。
+    ///     允许使用无参 <see cref="Action" /> 来订阅当前带参事件。
     /// </summary>
     /// <param name="onEvent">无参事件处理方法。</param>
-    /// <returns>IUnRegister 对象，用于稍后注销该事件监听器。</returns>
+    /// <returns><see cref="IUnRegister" /> 对象，用于稍后注销该事件监听器。</returns>
     IUnRegister IEvent.Register(Action onEvent)
     {
         return Register(Action);
 
-        void Action(T _, Tk __)
+        void Action(T _, TK __)
         {
             onEvent();
         }
@@ -104,8 +104,8 @@ public class Event<T, Tk> : IEvent
     ///     注册一个接受两个参数的事件监听器，并返回可用于取消注册的对象。
     /// </summary>
     /// <param name="onEvent">要注册的事件处理方法。</param>
-    /// <returns>IUnRegister 对象，用于稍后注销该事件监听器。</returns>
-    public IUnRegister Register(Action<T, Tk> onEvent)
+    /// <returns><see cref="IUnRegister" /> 对象，用于稍后注销该事件监听器。</returns>
+    public IUnRegister Register(Action<T, TK> onEvent)
     {
         _mOnEvent += onEvent;
         return new DefaultUnRegister(() => UnRegister(onEvent));
@@ -115,25 +115,25 @@ public class Event<T, Tk> : IEvent
     ///     取消指定的双参数事件监听器。
     /// </summary>
     /// <param name="onEvent">需要被注销的事件处理方法。</param>
-    public void UnRegister(Action<T, Tk> onEvent)
+    public void UnRegister(Action<T, TK> onEvent)
     {
         _mOnEvent -= onEvent;
     }
 
     /// <summary>
-    ///     触发所有已注册的事件处理程序，并传递参数 t 和 k。
+    ///     触发所有已注册的事件处理程序，并传递参数 <paramref name="t" /> 和 <paramref name="k" />。
     /// </summary>
     /// <param name="t">第一个参数。</param>
     /// <param name="k">第二个参数。</param>
-    public void Trigger(T t, Tk k)
+    public void Trigger(T t, TK k)
     {
         _mOnEvent?.Invoke(t, k);
     }
 
     /// <summary>
-    ///     获取当前已注册的监听器数量
+    ///     获取当前已注册的监听器数量。
     /// </summary>
-    /// <returns>监听器数量</returns>
+    /// <returns>监听器数量。</returns>
     public int GetListenerCount()
     {
         return _mOnEvent?.GetInvocationList().Length ?? 0;
