+++ GFramework.Core/Query/AbstractAsyncQuery.cs
@@ -1,5 +1,6 @@
 ﻿using GFramework.Core.Abstractions.Query;
 using GFramework.Core.Rule;
+using GFramework.Cqrs.Abstractions.Cqrs.Query;
 
 namespace GFramework.Core.Query;
 
@@ -24,4 +25,31 @@ public abstract class AbstractAsyncQuery<TResult> : ContextAwareBase, IAsyncQuer
     /// </summary>
     /// <returns>返回查询结果的异步任务</returns>
     protected abstract Task<TResult> OnDoAsync();
+}
+
+/// <summary>
+///     抽象异步查询基类，为需要输入参数的异步查询提供统一执行骨架。
+/// </summary>
+/// <typeparam name="TInput">查询输入类型，必须实现 <see cref="IQueryInput" /> 接口。</typeparam>
+/// <typeparam name="TResult">查询结果类型。</typeparam>
+/// <param name="input">查询输入参数。</param>
+public abstract class AbstractAsyncQuery<TInput, TResult>(TInput input)
+    : ContextAwareBase, IAsyncQuery<TResult>
+    where TInput : IQueryInput
+{
+    /// <summary>
+    ///     执行异步查询操作。
+    /// </summary>
+    /// <returns>返回查询结果的异步任务。</returns>
+    public Task<TResult> DoAsync()
+    {
+        return OnDoAsync(input);
+    }
+
+    /// <summary>
+    ///     抽象方法，用于实现具体的异步查询逻辑。
+    /// </summary>
+    /// <param name="input">查询输入参数。</param>
+    /// <returns>返回查询结果的异步任务。</returns>
+    protected abstract Task<TResult> OnDoAsync(TInput input);
 }
